# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/kiosks": {
      "post": {
        "operationId": "KiosksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKioskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kiosk created"
          },
          "403": {
            "description": "Caller is not an admin"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Create a kiosk (admin only, defaults to ACTIVE, no approval step)",
        "tags": [
          "Kiosks"
        ]
      },
      "get": {
        "operationId": "KiosksController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All kiosks across every tenant"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "List all kiosks (admin only)",
        "tags": [
          "Kiosks"
        ]
      }
    },
    "/kiosks/{id}": {
      "get": {
        "operationId": "KiosksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The kiosk"
          },
          "403": {
            "description": "Not this kiosk's owner and not an admin"
          },
          "404": {
            "description": "Kiosk not found"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get a kiosk by id (admin, or the owning kiosk-owner)",
        "tags": [
          "Kiosks"
        ]
      },
      "patch": {
        "operationId": "KiosksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKioskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kiosk updated"
          },
          "404": {
            "description": "Kiosk not found"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Update a kiosk, including revenueSharePct and status (admin only)",
        "tags": [
          "Kiosks"
        ]
      }
    },
    "/kiosks/{id}/status": {
      "patch": {
        "operationId": "KiosksController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateKioskStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated — takes effect on the kiosk's devices immediately"
          },
          "404": {
            "description": "Kiosk not found"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Toggle a kiosk between ACTIVE and INACTIVE (admin only)",
        "tags": [
          "Kiosks"
        ]
      }
    },
...
    "/locations": {
      "post": {
        "operationId": "LocationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Location created"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Create a location (admin specifies kioskId; kiosk-owner's own kiosk is implied)",
        "tags": [
          "Locations"
        ]
      },
      "get": {
        "description": "Admin sees all locations; kiosk-owner sees their own kiosk's locations; location-manager sees only their assigned locations.",
        "operationId": "LocationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Locations visible to the caller"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "List locations",
        "tags": [
          "Locations"
        ]
      }
    },
    "/locations/{id}": {
      "get": {
        "operationId": "LocationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The location"
          },
          "403": {
            "description": "Not in the caller's tenant scope"
          },
          "404": {
            "description": "Location not found"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Get a location by id",
        "tags": [
          "Locations"
        ]
      },
      "patch": {
        "operationId": "LocationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location updated"
          },
          "403": {
            "description": "Not in the caller's tenant scope"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Update a location, including tags",
        "tags": [
          "Locations"
        ]
      },
      "delete": {
        "operationId": "LocationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Location deleted"
          },
          "403": {
            "description": "Not allowed for this role or tenant"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Delete a location (admin or kiosk-owner only, not location-manager)",
        "tags": [
          "Locations"
        ]
      }
    },
    "/locations/{id}/setup-codes": {
      "post": {
        "operationId": "LocationsController_createSetupCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Setup code created"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Generate a new reusable device setup code for this location",
        "tags": [
          "Locations"
        ]
      },
      "get": {
        "operationId": "LocationsController_findSetupCodes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setup codes"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "List setup codes for this location",
        "tags": [
          "Locations"
        ]
      }
    },
    "/locations/{id}/setup-codes/{codeId}": {
      "patch": {
        "operationId": "LocationsController_updateSetupCode",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "codeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSetupCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setup code updated"
          },
          "404": {
            "description": "Setup code does not belong to this location"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "summary": "Revoke or reactivate a setup code",
        "tags": [
          "Locations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Saverlly API",
    "description": "Multi-tenant SaaS platform for internet kiosk businesses — admin console, kiosk portal, and machine (Chrome extension / desktop agent) endpoints.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "device-token": {
        "scheme": "bearer",
        "bearerFormat": "device-token",
        "type": "http",
        "description": "Opaque device token issued at /devices/register — used by machine clients only."
      }
    },
    "schemas": {
...
      "CreateKioskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Downtown Internet Cafe"
          },
          "revenueSharePct": {
            "type": "number",
            "example": 30,
            "description": "Percentage of commission the kiosk keeps (0-100)"
          },
          "contactEmail": {
            "type": "string",
            "example": "owner@kiosk.com"
          }
        },
        "required": [
          "name",
          "revenueSharePct",
          "contactEmail"
        ]
      },
      "UpdateKioskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "revenueSharePct": {
            "type": "number",
            "description": "Percentage of commission the kiosk keeps (0-100)"
          },
          "contactEmail": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        }
      },
      "UpdateKioskStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateLocationDto": {
        "type": "object",
        "properties": {
          "kioskId": {
            "type": "string",
            "description": "Required when the caller is ADMIN; ignored for KIOSK_OWNER (their own kioskId is always used)"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "tags": {
            "example": [
              "mall",
              "downtown",
              "high-traffic"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "address",
          "city",
          "state",
          "country"
        ]
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "tags": {
            "example": [
              "mall",
              "downtown",
              "high-traffic"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateSetupCodeDto": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "description": "Set false to revoke the code, true to reactivate it"
          }
        },
        "required": [
          "active"
        ]
      },
...
    }
  }
}